build: move loginSchema into schema folder
